[libgens] Ym2612: Ported over a few optimizations from Game_Music_Emu.
- Changed LFO_FMS_TAB[] from unsigned int to uint8_t. The largest
  value in the array is (LFO_FMS_BASE * 24) == ~24.6 == 24.

- Removed 'inline' from KEY_OFF() and KEY_ON(). These functions
  are too large for 'inline' to have any effect.

- Use ~0 instead of 0xFFFFFFFF.

- Optimized channel and slot number calculations in a few places.

- Optimized stereo channel masking.

- Added a few assert()s in reInit().

+++ src/libgens/sound/Ym2612_p.hpp
@@ -285,7 +285,7 @@ class Ym2612Private
 
 		static const uint8_t FKEY_TAB[16];
 		static const uint8_t LFO_AMS_TAB[4];
-		static const unsigned int LFO_FMS_TAB[8];
+		static const uint8_t LFO_FMS_TAB[8];
 
 		// Interpolation calculation.
 		int int_cnt;
